# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Authorbot API
  version: 1.0.0-alpha.0
  summary: Coordination-plane API for Authorbot book projects.
  description: |
    Skeleton of the Authorbot v1 API per design §15.

    Conventions (design §15.1):
    - All paths are prefixed with `/v1`.
    - Entity identifiers are UUIDv7, lowercase.
    - Timestamps are UTC RFC 3339 (`2026-07-19T18:00:00Z`).
    - Mutation endpoints require an `Idempotency-Key` header.
    - Editable resources return `ETag` and honor `If-Match`.
    - Errors are `application/problem+json` (RFC 9457).
    - List endpoints use cursor pagination (`cursor`, `limit`, `nextCursor`).
    - Every response carries `X-Correlation-Id`; asynchronous commands return
      an operation id resolvable at `/v1/projects/{projectId}/operations/{operationId}`.
    - Humans (GitHub-backed session cookie) and agents (bearer token) are
      treated identically after authorization.

    Implementation status: operations marked `x-implementation-status: planned`
    are design skeleton for later phases and are not served yet. Everything
    else matches the Phase 2 implementation (`apps/api`). Operations marked
    `x-dev-only: true` exist only when the deployment runs `AUTH_MODE=dev`;
    operations marked `x-auth-mode: github` only under `AUTH_MODE=github`.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Authorbot-Internal
servers:
  - url: /
    description: Same-origin deployment; paths already include the /v1 prefix.
security:
  - githubSession: []
  - agentToken: []
...
paths:
...
  /v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProject
      tags: [identity]
      summary: Return project metadata.
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/projects/{projectId}/publications:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listPublications
      tags: [operations]
      summary: List reported publications, newest first.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Publication history.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Publication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/publications:
    post:
      operationId: reportPublication
      tags: [operations]
      summary: Signed CI callback reporting build and deployment state.
      description: |
        Authenticated by `X-Authorbot-Signature-256`, an HMAC-SHA256 with
        WEBHOOK_SECRET over `<delivery-id>.<timestamp>.<raw-body>` — not by
        session or bearer credentials. The delivery id and timestamp are part
        of the signed material precisely because the delivery id is the replay
        suppression key: signing the body alone would let anyone holding one
        captured request resubmit it indefinitely under fresh delivery ids.

        `X-Authorbot-Timestamp` must be within 5 minutes of server time, so a
        captured request expires independently of the delivery ledger.
        Deliveries are additionally deduplicated on `X-Authorbot-Delivery`; a
        repeat delivery is a no-op regardless of its content.

        Example (bash):

            BODY='{"integratedCommit":"...","buildStatus":"succeeded"}'
            DELIVERY=$(uuidgen)
            TS=$(date -u +%Y-%m-%dT%H:%M:%SZ)
            SIG=$(printf '%s.%s.%s' "$DELIVERY" "$TS" "$BODY" \
              | openssl dgst -sha256 -hmac "$WEBHOOK_SECRET" -r | cut -d' ' -f1)

        This is the ONLY writer of publication state. Design §17.3: a revision
        is never marked published merely because its Git commit succeeded.

        A single commit's publication advances through several callbacks
        (`queued` → `building` → `succeeded` → deployed). Fields omitted from a
        callback keep their stored value; an explicit `null` clears one.
      security: []
      parameters:
        - name: X-Authorbot-Delivery
          in: header
          required: true
          description: Unique delivery id; replays are suppressed.
          schema:
            type: string
            maxLength: 200
        - name: X-Authorbot-Timestamp
          in: header
          required: true
          description: >
            RFC 3339 instant the callback was signed at. Must be within 5
            minutes of server time and is covered by the signature.
          schema:
            type: string
            maxLength: 64
        - name: X-Authorbot-Signature-256
          in: header
          required: true
          description: >
            `sha256=<hex hmac>` over `<delivery-id>.<timestamp>.<raw-body>`
            with WEBHOOK_SECRET.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [integratedCommit, buildStatus]
              additionalProperties: false
              properties:
                projectSlug:
                  type: string
                  description: >
                    Optional guard. When present it must name this deployment's
                    project, so a callback aimed elsewhere is rejected rather
                    than silently applied here.
                integratedCommit:
                  type: string
                  pattern: "^[0-9a-f]{7,64}$"
                buildStatus:
                  type: string
                  enum: [queued, building, succeeded, failed]
                deployedCommit:
                  oneOf:
                    - type: string
                      pattern: "^[0-9a-f]{7,64}$"
                    - type: "null"
                publicUrl:
                  description: Absolute http(s) URL only.
                  oneOf:
                    - type: string
                      format: uri
                    - type: "null"
                deployedAt:
                  oneOf:
                    - $ref: "#/components/schemas/Timestamp"
                    - type: "null"
                publisherVersion:
                  oneOf:
                    - type: string
                    - type: "null"
      responses:
        "200":
          description: Applied to an existing publication row, or recognized as a duplicate delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationCallbackResult"
        "201":
          description: First report for this integrated commit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationCallbackResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Missing or invalid callback signature.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    githubSession:
      type: apiKey
      in: cookie
      name: authorbot_session
      description: |
        Browser session established via GitHub OAuth login (design §19.1).
        HttpOnly, Secure, SameSite cookie; the value is an opaque session id.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Named, scoped, revocable, expiring agent token (design §19.2). Only a
        hash is stored server-side. Humans and agents are treated identically
        after authorization.
...
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: The project UUID or its slug (single-project deployment).
      schema:
        oneOf:
          - $ref: "#/components/schemas/Uuid"
          - $ref: "#/components/schemas/Slug"
...
  responses:
    BadRequest:
      description: Invalid payload.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid credential.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Actor lacks scope or role.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found or intentionally hidden.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Uuid:
      type: string
      format: uuid
      description: UUIDv7, lowercase.
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      examples:
        - 0190f301-7045-7b2d-9d91-95b3c8228b54
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 UTC timestamp.
      examples:
        - 2026-07-19T18:00:00Z
    Slug:
      type: string
      pattern: "^[a-z0-9][a-z0-9-]*$"
      description: Path-safe slug (no separators, no leading dots).
...
    Problem:
      type: object
      description: RFC 9457 problem details (`application/problem+json`).
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable machine-readable error code.
        correlationId:
          $ref: "#/components/schemas/Uuid"
      additionalProperties: true
...
    PublicationCallbackResult:
      type: object
      required: [duplicate, applied]
      properties:
        duplicate:
          type: boolean
          description: The delivery id had already been seen; nothing was applied.
        applied:
          type: boolean
        publicationId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        publication:
          oneOf:
            - $ref: "#/components/schemas/Publication"
            - type: "null"
    Publication:
      type: object
      description: >
        One integrated commit's publication state as reported by CI
        (design §17.3). Never inferred from a successful Git commit.
      required: [id, integratedCommit, buildStatus, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        integratedCommit:
          type: string
          description: Repository commit CI built from.
        buildStatus:
          type: string
          enum: [queued, building, succeeded, failed]
        deployedCommit:
          description: Commit actually serving the public site, per CI.
          oneOf:
            - type: string
            - type: "null"
        publicUrl:
          oneOf:
            - type: string
              format: uri
            - type: "null"
        deployedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        publisherVersion:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    PublicationStatus:
      type: object
      description: >
        Integrated versus deployed (design §17.3): `integratedCommit` is what
        Authorbot projected, `deployedCommit` is what CI says is live. They are
        reported separately so the gap is visible rather than assumed.
      required: [integratedCommit, deployedCommit, buildStatus, inSync]
      properties:
        integratedCommit:
          oneOf:
            - type: string
            - type: "null"
        deployedCommit:
          oneOf:
            - type: string
            - type: "null"
        buildStatus:
          description: Build status of the most recent publication CI reported.
          oneOf:
            - type: string
              enum: [queued, building, succeeded, failed]
            - type: "null"
        publicUrl:
          oneOf:
            - type: string
              format: uri
            - type: "null"
        deployedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        publisherVersion:
          oneOf:
            - type: string
            - type: "null"
        inSync:
          type: boolean
          description: >
            True only when both commits are known AND equal. An unknown
            deployment is never reported as up to date.
        latest:
          oneOf:
            - $ref: "#/components/schemas/Publication"
            - type: "null"
    Project:
      type: object
      required: [id, slug, repoProvider, repo, defaultBranch, status, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        slug:
          $ref: "#/components/schemas/Slug"
        repoProvider:
          type: string
          enum: [github]
        repo:
          type: string
          description: "`owner/name`, e.g. `JoeMattie/causal-projector`."
        defaultBranch:
          type: string
        status:
          type: string
          enum: [active, invalid, diverged, archived]
        gitIntegration:
          type: string
          enum: [configured, unconfigured, incomplete, invalid]
          description: >
            Whether this deployment holds a usable GitHub App credential set
            (Phase 5 §2). `unconfigured` means reads work and mutations are
            accepted and queued, but nothing is committed to the repository;
            `incomplete` means some but not all of GITHUB_APP_ID,
            GITHUB_APP_PRIVATE_KEY and GITHUB_INSTALLATION_ID are set;
            `invalid` means all three are present but at least one cannot work
            (a PKCS#1 private key, a non-numeric app or installation id).
            `incomplete` and `invalid` are both treated exactly like
            `unconfigured` at runtime. `configured` means all three are present
            and well-formed. No credential value is ever exposed.
        projection:
          type: object
          description: >
            State of the database projection relative to the repository
            (Phase 5 §6, design §20.3).
          required: [commit, stale]
          properties:
            commit:
              description: >
                Repository commit the projection was last built from — the
                *integrated* commit. Null before the first successful refresh.
              oneOf:
                - type: string
                - type: "null"
            stale:
              type: boolean
              description: >
                A verified push has arrived that the projection has not yet
                absorbed. Set before the refresh is attempted, so a refresh
                that never ran is still visible as owed work.
        divergence:
          type: object
          description: >
            Design §14.5 divergence state. While `diverged`, prose writes are
            refused with the `project-diverged` problem and reads continue.
          required: [state]
          properties:
            state:
              type: string
              enum: [ok, diverged]
            divergedAt:
              oneOf:
                - $ref: "#/components/schemas/Timestamp"
                - type: "null"
            kinds:
              type: array
              items:
                type: string
                enum: [revision-regressed, anchor-blocks-vanished]
            chapters:
              type: array
              items:
                type: object
                properties:
                  chapterId:
                    $ref: "#/components/schemas/Uuid"
                  path:
                    type: string
                  kind:
                    type: string
                  projectedRevision:
                    type: integer
                  snapshotRevision:
                    type: integer
        publication:
          $ref: "#/components/schemas/PublicationStatus"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...